Better Scoring Algorithm (closes #16)

+++ Desktop.h
@@ -17,7 +17,7 @@ class Desktop {
         std::string exec;
         int icon_type = ICON_TYPE_NAME;
         int is_valid = 0;
-        int priority = 0;
+        float priority = 0;
         explicit Desktop(std::string file_path);
         bool exec_to_launch = false;
 };
